# perf(scoring/dag): optimize `get_labels`